moved global hotkeys into their own crate
