+++ gonk-hotkeys/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "gonk-hotkeys"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+win-hotkey = {git = "https://github.com/zX3no/win-hotkey"}
+gonk-database = { path = "../gonk-database" }
+gonk-server = {path = "../gonk-server"}
