+++ gonk/Cargo.toml
@@ -16,6 +16,3 @@ static_init = "1.0.2"
 unicode-width = "0.1.9"
 optick = "1.3.4"
 gonk-server = {path = "../gonk-server"}
-
-[target.'cfg(windows)'.dependencies]
-win-hotkey = { git = "https://github.com/zX3no/win-hotkey", rev = "1d6cb22ba202a445f73bf51a7e785655440639b3" }
